Rename file for open bank account domain command
