+++ source/imports/domain/bank-accounts/bank-account.app.tests.js
@@ -3,7 +3,7 @@ import { mocha } from 'meteor/avital:mocha';
 // import { chai, assert } from "meteor/practicalmeteor:chai";
 import BankingDomain from '../banking-domain-module';
 import BankAccount from './bank-account';
-import OpenBankAccount from '../../domain-commands/open-account-command';
+import OpenBankAccount from '../../domain-commands/open-bank-account';
 import BankAccountOpened from '../../domain-events/bank-account-opened';
 import Contact from '../../value-objects/contact';
 import NIN from '../../value-objects/nin';
