+++ source/imports/es-routers/bank-account-router.js
@@ -1,5 +1,5 @@
 import BankAccount from '../domain/bank-accounts/bank-account';
-import OpenBankAccount from '../domain-commands/open-account-command';
+import OpenBankAccount from '../domain-commands/open-bank-account';
 
 const BankAccountRouter = Space.eventSourcing.Router.extend('BankAccountRouter', {
